Create Day 2 - Tip Calculator
The challenge at the end of my learning module was to utilize the material learned to create a calculator that can perform basic arithmetic to output the total amount spent on a particular meal. This can be further broken down into a split between any number of people.

Again, this is just for learning purposes.